Base system struct
